Add JsonPath and classes to retrieve author of revisions.
JsonPath was added to build.gradle.

RevisionParser was created in the model folder. Its job as of now is to get
the author of the most recent revision of the wikipedia article entered.

RevisionParserTest was added to ensure that RevisionParser is able to parse
the json output of the "Soup" wiki page, which is stored locally.

WikipediaRevisionApplication had a new method, getLatestRevisionOf, added
and as of now includes the code to construct a URL which leads to
Wikipedia's API.